requirements.txt
langchain==0.3.27
langchain_community==0.3.29
langchain_core==0.3.75
langchain_google_genai==2.1.10
langchain_huggingface==0.3.1
langchain_openai==0.3.32
langgraph==0.6.6
python-dotenv==1.1.1
typing_extensions==4.15.0
